strip HTML tags from description string
