+++ tests/unit/utils/test_html.py
@@ -1,0 +1,16 @@
+"""Tests for the utils.html module."""
+
+import pytest
+
+from nipoppy.utils.html import strip_html_tags
+
+
+@pytest.mark.parametrize(
+    "html,tags,expected",
+    [
+        ("<div><p>Test</p></div>", None, "Test"),
+        ("<div><p>Test</p></div>", {"p"}, "<div>Test</div>"),
+    ],
+)
+def test_strip_html_tags(html, tags, expected):
+    assert strip_html_tags(html, tags) == expected
